requirements.txt
emoji
python-dotenv
psycopg2